Avances

+++ 3D Terrain Generator/src/Heightmap/Heightmap.cpp
@@ -16,7 +16,7 @@ Heightmap::Heightmap(int width, int height):_exponent(0), _width(width), _height
 
 }
 
-Heightmap::Heightmap(int exponent):_exponent(exponent), _width(pow(2,_exponent)), _height(_width), _heightmap(_width, vector<float>(_height))
+Heightmap::Heightmap(int exponent):_exponent(exponent), _width(static_cast<int>(pow(2,_exponent))), _height(_width), _heightmap(_width, vector<float>(_height))
 {
 	
 }
@@ -48,7 +48,7 @@ void Heightmap::Resize(int width, int height)
 void Heightmap::Resize(int exponent)
 {
 	_exponent = exponent;
-	_width = pow(2, exponent);
+	_width = static_cast<int>(pow(2, exponent));
 	_height = _width;
 	Resize(_height, _width);
 }
